requirements.txt
# ── Flask & Extensions ────────────────────────
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-jwt-extended>=4.6.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
gunicorn>=21.2.0

# ── Database (PostgreSQL for production) ──────
psycopg2-binary>=2.9.9

# ── PyTorch & ML ──────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
safetensors>=0.4.0

# ── Computer Vision ──────────────────────────
opencv-python>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0

# ── Face Detection ────────────────────────────
facenet-pytorch>=2.5.3
